refactor: 결제 요청, 결제 트랜잭션 조회 예외 처리 추가 및 코드 리팩토링 - 결제 요청, 결제 트랜잭션 조회 fallback 처리 수정 - 결제 요청, 결제 트랜잭션 조회 dto 경계 나누는 코드 추가

+++ apps/commerce-api/src/test/java/com/loopers/domain/order/OrderAppServiceIntegrationTest.java
@@ -66,7 +66,7 @@ public class OrderAppServiceIntegrationTest {
             pointFixture.save(user.getId(), balance);
 
             Long productPrice = 100L;
-            Integer stockQuantity = 1;
+            Long stockQuantity = 1L;
             Product saveProduct = productWithStockFixture.save(
                     brandFixture.save(),
                     productPrice,
