+++ apps/commerce-api/src/test/java/com/loopers/fixture/product/ProductWithStockFixture.java
@@ -4,5 +4,5 @@ import com.loopers.domain.brand.Brand;
 import com.loopers.domain.product.Product;
 
 public interface ProductWithStockFixture {
-    Product save(Brand brand, Long productPrice, Integer stockQuantity);
+    Product save(Brand brand, Long productPrice, Long stockQuantity);
 }
